requirements.txt
flask
flask_cors
pandas
matplotlib
azure.cognitiveservices.speech
pydub
tiktoken
pymysql
google-api-python-client
google-auth-oauthlib
python-pptx
openai